requirements.txt
feedparser>=6.0.10
beautifulsoup4>=4.12.0
markdown>=3.5.0
pyyaml>=6.0
yfinance>=0.2.40
